secure_flask_app/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
python-dotenv
gunicorn
pytest
